+++ source/ProcessManager.Orchestrations/Processes/BRS_026/V1/Activities/EnqueueRejectMessageActivity_Brs_026_V1.cs
@@ -12,12 +12,13 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+using Energinet.DataHub.ProcessManager.Abstractions.Components.BusinessValidation;
+using Energinet.DataHub.ProcessManager.Components.BusinessValidation;
 using Energinet.DataHub.ProcessManager.Components.EnqueueActorMessages;
 using Energinet.DataHub.ProcessManager.Core.Application.Orchestration;
 using Energinet.DataHub.ProcessManager.Core.Domain.OrchestrationInstance;
 using Energinet.DataHub.ProcessManager.Orchestrations.Abstractions.Processes.BRS_026.V1.Model;
 using Microsoft.Azure.Functions.Worker;
-using NodaTime;
 
 namespace Energinet.DataHub.ProcessManager.Orchestrations.Processes.BRS_026.V1.Activities;
 
@@ -25,11 +26,9 @@ namespace Energinet.DataHub.ProcessManager.Orchestrations.Processes.BRS_026.V1.A
 /// Enqueue reject message in EDI (and set step to running)
 /// </summary>
 internal class EnqueueRejectMessageActivity_Brs_026_V1(
-    IClock clock,
     IOrchestrationInstanceProgressRepository progressRepository,
     IEnqueueActorMessagesClient enqueueActorMessagesClient)
 {
-    private readonly IClock _clock = clock;
     private readonly IOrchestrationInstanceProgressRepository _progressRepository = progressRepository;
     private readonly IEnqueueActorMessagesClient _enqueueActorMessagesClient = enqueueActorMessagesClient;
 
@@ -41,27 +40,29 @@ internal class EnqueueRejectMessageActivity_Brs_026_V1(
             .GetAsync(input.InstanceId)
             .ConfigureAwait(false);
 
-        orchestrationInstance.TransitionStepToRunning(
-            Orchestration_Brs_026_V1.EnqueueActorMessagesStepSequence,
-            _clock);
-        await _progressRepository.UnitOfWork.CommitAsync().ConfigureAwait(false);
-
         await EnqueueRejectMessageAsync(orchestrationInstance.Lifecycle.CreatedBy.Value, input).ConfigureAwait(false);
     }
 
     private Task EnqueueRejectMessageAsync(OperatingIdentity orchestrationCreatedBy, ActivityInput input)
     {
         // TODO: Set correct data when async validation is implemented
+        var rejectedMessage = new RequestCalculatedEnergyTimeSeriesRejectedV1(
+            ValidationErrors: input.ValidationErrors
+                .Select(e => new ValidationErrorDto(
+                    Message: e.Message,
+                    ErrorCode: e.ErrorCode))
+                .ToList());
+
         return _enqueueActorMessagesClient.EnqueueAsync(
             Orchestration_Brs_026_V1.UniqueName,
             input.InstanceId.Value,
             orchestrationCreatedBy.ToDto(),
             input.IdempotencyKey,
-            input.RejectedData);
+            rejectedMessage);
     }
 
     public record ActivityInput(
         OrchestrationInstanceId InstanceId,
-        RequestCalculatedEnergyTimeSeriesRejectedV1 RejectedData,
+        IReadOnlyCollection<ValidationError> ValidationErrors,
         Guid IdempotencyKey);
 }